Revert back to voltage flywheel

+++ include/flywheel.hpp
@@ -13,4 +13,4 @@ void setflypct(int set);
 
 void flywheelControl();
 
-int getFly();
+double getFly();
